Increasing size of guild logo field

+++ datadrivendota/guilds/models.py
@@ -4,4 +4,4 @@ from django.db import models
 class Guild(models.Model):
     steam_id = models.IntegerField()
     name = models.CharField(max_length=50)
-    logo = models.IntegerField()
+    logo = models.BigIntegerField()
